requirements.txt
# test
pytest
pytest-asyncio

aiofiles

click

werkzeug

hypercorn
flask[async]>=3.1.3
asgiref

python-socketio
dotenv
pdfplumber

aioboto3
botocore


asyncpg
redis
sqlparse

pyjwt
flask-jwt-extended

httpx
thefuzz

git+https://github.com/boris-kanga/kb_tools.git#egg=kb_tools

